requirements.txt
rich
dnspython